refactor: config를 통해 UseCase 생성하도록 수정
