+++ src/main/java/menu/Application.java
@@ -1,21 +1,17 @@
 package menu;
 
-import menu.common.ConsoleLogger;
-import menu.common.Logger;
+import menu.common.config.MenuRecommendConfig;
+import menu.common.log.ConsoleLogger;
+import menu.common.log.Logger;
 import menu.controller.MenuRecommendController;
-import menu.service.MenuRecommend;
-import menu.service.SelectCategory;
 import menu.step.RecommendStartStep;
 import menu.step.Step;
-import menu.usecase.MenuRecommendUseCase;
-import menu.usecase.SelectCategoryUseCase;
 
 public class Application {
     public static void main(String[] args) {
-        MenuRecommendUseCase menuRecommend = new MenuRecommend();
-        SelectCategoryUseCase selectCategoryUseCase = new SelectCategory();
+        MenuRecommendConfig config = new MenuRecommendConfig();
         MenuRecommendController controller =
-                new MenuRecommendController(menuRecommend, selectCategoryUseCase);
+                new MenuRecommendController(config);
         Logger consoleLogger = new ConsoleLogger();
         Step step = new RecommendStartStep(controller, consoleLogger);
